Readme
First, a function accepts four inputs: the first index, last index, list, and to_search (item to be searched).
Then we check if the value of the last index is greater than or equal to the value of the first index. If the condition is true, we assign the operation of finding the middle element index to the variable name "mid_index". Then the middle element is gotten from the list using the mid index as position.
We create an "if" statement under the first "if" block to check if the middle element and the variable "to_search" are equal. If they are equal, the position of the item will be returned.
Then we check for the second condition, (if middle element != item to be searched) which leads us to two scenarios:
– if the middle element is greater than the item to be searched, the new position will shift to the left once. The search will begin from the first index and end at the new position. We return the function and pass in the new position as the last index value.
– if the middle element is less than the item to be searched, the new position will shift to the right once. The search will begin from the new position and end at the last index. We return the function and pass in the new position as the first index value.
The last condition will be on the same indent as the first "if" statement. If the to_search is not in the list, it will return a statement
The final step is to call the function and the result is displayed.